+++ app/mailers/devise_sendgrid_mailer.rb
@@ -3,14 +3,15 @@ class DeviseSendgridMailer < Devise::Mailer
   include Devise::Controllers::UrlHelpers
   include Rails.application.routes.url_helpers
 
-  def confirmation_instructions(record, token, opts={})
-    link_url = api_ui_v1_user_confirmation_url(confirmation_token: token)
-    SendgridMailer.account_confirmation(record, link_url: link_url).deliver
+  def confirmation_instructions(user, token, opts={})
+    hostname = get_user_domain(user)
+    link_url = api_ui_v1_user_confirmation_url(confirmation_token: token, host: hostname)
+    SendgridMailer.account_confirmation(user, link_url: link_url).deliver
   end
 
-  def reset_password_instructions(record, token, opts={})
-    link_url = edit_api_ui_v1_user_password_url(reset_password_token: token)
-    SendgridMailer.reset_password(record, link_url: link_url).deliver
+  def reset_password_instructions(user, token, opts={})
+    hostname = get_user_domain(user)
+    link_url = edit_api_ui_v1_user_password_url(reset_password_token: token, host: hostname)
+    SendgridMailer.reset_password(user, link_url: link_url).deliver
   end
-
 end